requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
alembic
python-multipart
reportlab